Refactor extractor Broke up extractor into config and exporter. Exporter now main function.
